+++ CHANGELOG.md
@@ -9,9 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 * Support for setting path to service key file in config
+* Ability to translate into multiple comma-separated languages
 
 ### Fixed
-* Now fails gracefully is path to key JSON is missing or invalid
+* Now fails gracefully if path to key JSON is missing or invalid
 
 
 ## [0.3.0] - 2019-06-20